general findings

+++ server/src/main/java/dev/findfirst/bookmarkit/users/model/user/Role.java
@@ -18,7 +18,7 @@ import lombok.NoArgsConstructor;
 public class Role {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer role_id;
+  private Long role_id;
 
   @Enumerated(EnumType.STRING)
   @Column(length = 20)
@@ -28,11 +28,11 @@ public class Role {
     this.name = name;
   }
 
-  public Integer getId() {
+  public long getId() {
     return role_id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.role_id = id;
   }
